Create README.md
Creating a Roblox game knowing only the minimum amount of programming and even on a cell phone! But of course, with the help of a person with a computer but also with minimal knowledge of code.
I'll be here to get tips on Roblox programming and its functions, in addition to making a game (my dream) I'll be gaining more knowledge of LUA codes!